[kopenwithdialog] Show better error message when specifying a non-executable file
If the user uses the file picker to specify a file instead of an
application the file has to be executable.

If it isn't "$file not found, please type a valid program name" is
shown, which is confusing. Instead show that the file is not executable
so the user knows what's going on.

BUG: 437880