KMountPoint: add a method to get the mount Id (util-linux/libmount specific)
The mount Id is the first field in /proc/self/mountinfo, and it's unique per
mount (the value is only reused if the the device is unmounted).

This is useful to differentiate between a base/regular mount point and a
bind mount; this is a step towards making kio_trash handle bind mounts, so
that deleting files from a bind mount doesn't always move the files to
~/.local/share/Trash.

Tweak API docs a bit.