KMountPoint: match by device ID (dev_t) in findByPath()
- Add a method to get the device ID for a mount point; libmount has API to
  do that, and for other code paths use stat()

- Match by device ID in findByPath(), if the partition is mounted we'll get
  its device ID, and if it isn't mounted, the device ID of e.g.
  /mnt/unmounted-part/ will be that of the partition where "/" is mounted
  (i.e. where the dir /mnt/unmounted-part/ exists)

Each mount is supposed to have a unique device ID/major/minor (even for
network mounts).

For a bind mount, e.g. /home/user/bar, the device id is the device id of the
the base mount point, e.g. /mnt/foo; to make sure we get the actual mount
point, check that the "path" we're trying to find a KMountPoint for, starts
with mountPoint(). This is slightly better than what the original code did
(comparing the paths lenght, till it found the longest matching path), since
half of the identifying now is done with the device id, which is deterministic
most of the time... except for bind mounts :)