kfilepropsplugin: add a .ui file
Ui-wise it looks the same.

- The uic creates all widgets, then we show/hide the appropriate ones as
needed (e.g. there is a KIconButton (to be able to change the icon) and an
icon QLabel (when the icon can't be changed)). In the code one can check
isVisibleTo() to find which one is used ...etc

- Move the UDSEntry extra fields to the bottom of the layout, since we don't
know how many rows we'll need at build time, I couldn't fit that in a .ui
file; so appending to the bottom of the layout is the simplest way