kfilepropsplugins: remove some no-op code
This is supposed to select the filename without selecting the extension, in
the filename line edit, but it looks like it hasn't been working worked for
some time. My guess is that the setFocus() call is overriden somewhere else
in the vast hierarchy of widgets in the dialog.

NO_CHANGELOG