Change KIO's default behavior to fork slaves rather than use klauncher
Several developers have tested this for a few months, and it all seems
to be working fine.

The use case for starting slaves via klauncher was for a process to
start a slave, put it on hold, and pass it to another process.
With the absence of KIO-based webbrowsers, and the removal of the
option that was actually doing this for HTTP urls ("use application
depending on file type"), this use case has disappeared.