kio_trash: set an extra UDSEntry field to hold the actual path of an item
This way when e.g. the KPropertiesDialog is called on an item in trash:/,
we can show the physical path to the trashed items (e.g.
~/.local/share/Trash/files or /path/to-other-mount-point/.Trash-$UID/files/).
This gives the user a clue where the files reside on the disk, instead of
just "/ (trash)", trash:/ is a virtual view on all the trash directories.

Notes (thanks to dfaure for pointing them out):
  - setting UDS_URL wouldn't work here, as it would mess up the tree view
    hierarchy in trash:/ (e.g. in Dolphin), (this was mentioned in a commit
    in kdelibs)
  - setting UDS_LOCAL_PATH would break DnD'ing stuff to/from trash:/ among
    other stuff

BUG: 282919
FIXED-IN: 5.85