kio_trash: set UDS_LOCAL_PATH for entries in trash:/
This way when e.g. the KPropertiesDialog is called on an item in trash:/,
the Location: can be set to the actual path to the item (e.g.
~/.local/share/Trash/files or /path/to-other-mount-point/.Trash-$UID/files/).
This gives the user a clue where the files reside on the disk, instead of
just "/ (trash)", trash:/ is a virtual view on all the trash directories.