KDirOperator: expand to url only in detail tree view
When "allow expansion" is disabled in the menu, the view type is KFile::Tree,
however when selecting an item from the location (Name:) combobox history,
the item model would expand to show the file; that should only happen if
"allow expansion" is enabled, i.e. when the view type is KFile::DetailTree.

BUG: 440475
FIXED-IN: 5.86