Deprecate the last bit of QDomDocument in the public KIO API
This replaces the QDomDocument DavJob::reponse() with a method returning
the raw response byte array instead, leaving the choice of the XML parser
to use to the user.

With this being the last use, Qt::Xml is now only used internally and thus
can be moved to the private link interface (where it wrongly is already).