Add icons for XDG publicshare and templates dirs
In the home folder for new users the directories "Public" and "Templates"
stand out, because they do not have a special icon.
Icons for them are available in Breeze, but not used as Qt does not implement
them as StandardPaths (QTBUG-86106 and QTBUG-78092).
The change here adds icons for both folders by looking up the XDG directories
and adding them to the map of standard locations.