Exclude various non-functional components from the Android build
These are mainly out-of-process components and/or things heavily relying
on D-Bus, things not available/unreachable on Android, and thus this should
not impact e.g. Okular on Android.

The main objective of this change is to focus subsequent D-Bus removal
patches on the components we actually need building on Android.