Fix KDirModelTest::testChmodDirectory() unit test
Changing the permissions of a directory in a KDirModel, and calling
KDirNotify::emitFilesChanged() didn't cause the item in the KCoreDirLister
to get refreshed, following the code:
KCoreDirListerCache::slotFilesChanged()
KCoreDirListerCache::processPendingUpdates()
KCoreDirListerCache::emitRefreshItem()
KCoreDirListerPrivate::addRefreshItem()

isItemVisible() would return false for both the old and new KFileItems'
since this is the root item ".".

This unit test started failing after the changes in KMountPoint, which led
to changing the temp dirs used by KCoreDirLister and KDirModel from /tmp
(which is usually mounted as tmpfs nowadays) to a regular directory in
$HOME/.qttest/.

This commit makes sense anyway (the root item of a kdirmodel is always
"visible").

NO_CHANGELOG