Build on Android without D-Bus
This is the big ugly rest. This is mostly disabling various bits of code
that are non-functional (due to lack of D-Bus) or not used by Okular
(the only active KIO user on Android). For ifdef'ing it uses a dedicated
define to make clear this isn't meant to be final (unlike previous patches
that removed concepts simply not applicable on Android).

This isn't making things worse compared to having the removed code call
into D-Bus stubs instead. It however makes KIO (and thus Okular) build in
our regular Android development environment without the need for the D-Bus
stub, thus making further work on this much easier.

# Conflicts:
#	CMakeLists.txt