KDirOperator: provide an option to enable showing open-with item actions
The use case here is KFileWidget, even though it sounds weird to add
open-with actions to the open file dialog, it still makes sense if you're
uploading an image to some web site, while picking the image you may want
to open it in a full-fledged viewer; another example, when uploading/attaching
a text file to a bug report or mail client composer, you may want to check
the contents of the file. This is more intuitive than 'open containing folder'
then opening the file.

BUG: 440748
FIXED-IN: 5.87