file_unix: minor cleanup
Separate ::chmod() from acl_set_file() call, more readable that way.

If the permissions are -1, that means preserve the permissions/attributes
of the source file, this matches what the code is/was doing and the docs.
Maybe we should add a couple more setters for the job, e.g. setPermissions(),
setAcl(), then users can fine-tune the permissions of the dest, preserve,
don't set, leave as default

CCBUG: 103331