DBus activation with filemanager iface
In Dolphin on wayland currently, if you right clicks a file and create a
zip file from it, Dolphin makes a new window.
What we want to happen is Dolphin to focus the window we have with the
file selected.

This patches KIO::highlightInFileManager() to call
KWindowSystem::requestXdgActivationToken(..) and set the result as the
OpenFileManagerWindowJob's startupID