KPropertiesDialog: don't block the parent app while "Advanced Options" is open
By making the advanced options dialog non-modal, and disabling the button
that launches it in the parent KPropertiesDialog while it's open, so that
you can't open multiple instances of the advanced options dialog.

The QDialog::finished() signal is not emitted by the "advanced options" dialog,
if the parent KPropertiedDialog is closed while the former is still open;
and closing the KPropertiesDialog causes the advanced one to get closed, so
no races that I can see.

BUG: 439900
FIXED-IN: 5.86