Update KUrlRequester::setMode() api docs
The underlying file dialog is created with mode:
"KFile::File | KFile::ExistingOnly | KFile::LocalOnly" by default, which is
what "fileDialogMode" is set to in KUrlRequesterPrivate constructor initializer
list.

Thanks to yecril71pl for pointing the issue out on IRC #kde-devel.

Thanks to Friedrich for pointing out where the default is actually set, I
was blind it seems.

GIT_SILENT