CopyJob: handle the no-op case of symlinks on FAT partitions
This copies some code from file_unix (see commit c27196acc33243db1da) to
CopyJob, since the latter is a more central place, we can catch more symlink
operations.

And as noted by dfaure in the MR, the code in file_unix is still needed for
the use case of apps using FileCopyJob directly and not CopyJob.

Split the msdos filesystem related code to a separate method, more readable
this way, I hope.

BUG: 442310
FIXED-IN: 5.88