Remove cmake option KIO_FORK_SLAVES
Developers with an old CMakeCache.txt could still end up using
klauncher, while the rest of the world is now forking slaves.

Removing forkSlaves() led to removing klauncher()->requestSlave()
as well as emptying the methods that were already disabled by
forkSlaves() being always true, such as holdSlave() and
checkForHeldSlave(). This, in turn, meant m_checkOnHold (in JobPrivate
and in SchedulerPrivate) were obsolete, removed them too.

Only dead code was harmed in the process of this commit.