Allow loading of Thumbnailers using embedded json metadata
This way thumbnailer plugins don't need to be loaded using the KServiceTypeTrader
mechanism, but can use embedded json metadata.
KServiceTypeTrader is planned to go away in KF6.

This also simplifies the code in kio-extras since we can then load the metadata
from the binary path and don't have to use KServiceTypeTrader to get the desktop file
for the given binary.

By using KPluginInfo to convert the metadata to json on the fly, we don't need to
do another query for the file on disk and then a reparsing of the desktop file.
This is ~2.5 times slower and even requires a few more lines to implement.

Task: https://phabricator.kde.org/T12179