[KPropertiesDialog] Improve presentation of file type configuration
Instead of showing a big "Configure File Type" or "Create File Type" button,
show which application will open this file by default with an option
to change that.

The logic when clicking the dialog (editing existing file type or
creating a new one) is unchanged but the visual representation is
made more self-explanatory.

While at it, fix KAuthorized check for `editfiletype` not working.