De-duplicate service menus from KServiceTypeTrader and KFileUtils results
In case one has installed an older version of an app and a local build
which used the new install location the results will be duplicated.
Same if one just does `make install` after updating the app - in which
case the old file from disk will not be removed.

Relates to https://phabricator.kde.org/T14543