[knewfilemenu] Replace setPopupFiles with setWorkingDirectory
For all intents and purposes setPopupFiles is the current directory, i.e. new files are created relative to it.

The name doesn't really reflect that.

It being a list is also weird since we are only really interested in passing one file. Some parts of the implementation only look at the first element in the list