[KFilePlacesView] Add middle click support and signals for item activation
This adds signals for `placeActivated`, `tabRequested`, `activeTabRequested`,
and `newWindowRequested`, similar to how it was done in `KUrlNavigator`.

The signals are emitted after mounting a storage device, if it needs to be
mounted first. The regular `clicked` signal would be emitted immediately
after clicking, where the storage URL might not be known yet.

Also, add "Open in New Tab" and "Open in New Window" to the context menu.