[KFilePlacesView] Wire up QAbstractItemView::iconSize
Instead of tracking it purely internally and in the delegate.
This way an application can call `setIconSize` and have this change
actually persisted as well as query the current iconSize.

Using an empty `QSize` for "automatic" is somewhat questionable
but the view never really communicated its icon size to the
outside world, so nobody should have been relying on it being accurate.