Add PreviewJob::availableThumbnailerPlugins method
Dolphin only uses KServiceTypeTrader for checking the available plugins, which has been deprecated.
For the json based thumbnailers an extra code path would need to be added, but by providing a getter
for the available plugins we keep the compat logic in one place. Also one does not need to adjust the code when the
plugins get ported to the "kf6" plugin namespace.