file_unix: Do not try to preserve ownership when permission is -1
When the permission mode is set to -1, it means the copy job doesn't want to
touch the dest permissions, but the dest ownership is always touched, which
results in root password being required when creating a new file from
templates.

This commit makes the copy job ignore dest ownership when the permission
mode is -1, so no authentication dialog will pop up when creating a new
file from templates.

BUG: 447779