Install kio_version.h in the KIO include prefix
See https://mail.kde.org/pipermail/kde-frameworks-devel/2022-January/120608.html
for the background on this.

This is considered compatible as both prefixes are included in the build
metadata, and local tests with Plasma and PIM confirm that so far.

I picked KIO as a test subject as kio_version.h is one of the most widely
used version headers. I'd suggest we try this here for 5.91 and if that
doesn't cause any fallout roll this out to all frameworks.