Emergency deletion of trashFileOwnedByRoot test
This test tried to move /etc/paswd into the trash which was an
incredibly dangerous thing to try to do, because if that move
operation ever succeeded, it would brick your whole system!

With polkit support being merged, this test now does just that every
time it's run.

This commit is an emergency removal so that people don't brick their
systems by running KIO's test suite anymore.