Remove unneeded json metadata for kiod plugins
Unlike the other kded modules, the affected plugins by this commit are installed in the "kf5/kiod" namespace.
Those are considered KIO-internal and do not get displayed in the "Background Services" KCM.
Consequently we do not need the name or description. Keeping them around would only increase build
times and binary sizes. Also the translators might put effort into sth. that is not visible to any users.

The ServiceTypes are unused and consequently removed.