[desktopexecparser] Consider associations from mimeapps.list to determine whether an app supports a scheme
Currently we check whether an app has MimeType=x-scheme-handler/foo in its desktop file.

This breaks for some apps, e.g. thunderbird handles mailto but does not always have it marked as such in its desktop file.

Also it ignores manual associations in mimeapps.list

Check whether the relevant app is in the list of possible handlers for that url scheme and don't use kioexec in that case

BUG: 440062
BUG: 439132