Introduce API to unwatch directories
If a client calls listDir(someUrl, keep=true) kdirwatcher adds the
listdir to both: ` listersCurrentlyListing` and
`listersCurrentlyHolding`. This is typically used by treeviews.

When we call stop(someUrl) we only remove it from
listersCurrentlyListing not  listersCurrentlyHolding. This means any
file changes in that URL will still be emitted.

As well as being wasteful this causes a user facing issue in dolphin
where it gets an entry is emitted for which it cannot find the parent in
it's local model. Dolphin gets confused and inserts it at the root.

This introduces an API that we can use to stop change notifications for
a given directory.