Split KIO::DesktopExecParser unit tests out of krununittest
The plan is to move DesktopExecParser to a KF module in a higher tier
(e.g. KService), so splitting the unit tests now is one step closer to
that goal; and it's clear now that it's not tied to KRun, so no point
putting its unit tests within krununittest.

GIT_SILENT