Draft: Add persistent file ID support to KIO
This adds support for using the various OS facilities for getting
a persistent file ID to identify a file regardless of moving/renaming.

This does not add support for using a persistent file ID to open/retrieve
a file, since that needs elevated permissions on Linux and BSD to perform.