fix reading wrong integer type from config
KConfig uses the default value to derive the type of the setting.
This forces KConfig to return the correct integer type KIO::filesize_t
instead of using plain 0, which is probably just an int.
KConfig interpreting a value to big for an int or uint as such
seems to be the cause of the following issue.

When MaximumRemoteSize was set to 4096 MiB or more,
previews of remote files were not being shown,
since dolphin believed the limit to be 0 when it was set to 4096 MiB.

before:

Setting 4095 MiB / 4293918720 Byte was read as 18446744073708503040 Bytes
Setting 4096 MiB / 4294967296 Byte was read as 0 Bytes
Setting 4097 MiB / 4296015872 Byte was read as 1048576 Bytes

These values are produced by interpreting the config value as an int,
which overflows if the value is large enough.
That int is then stored into a qulonlong again,
converting the overflowed signed int into an unsigned long long.

Indicating the type we actually want to KConfig avoids all of the above
and the values are now read / interpreted correctly and as expected.

BUG: 452048