Don't put job/ioworker on hold after getting the mimetype
we only call get() so we can get the mimetype, we are in fact not
interested in the data. so, kill the job instead of putting it on hold.
putting it on hold means it will forever sit there waiting for
resumption. what's more it still holds any opened resources during this
time frame, which, among other things, includes file locks when used
with smb://

(kill is implicitly executed when the parent job ends)

BUG: 452729