KFileWidget: KF6, fix view size glitch on initial show
When the open file dialog is first shown, the icon/details...etc view is
very small in size, taking a corner of the KDirOperator view; switching to
another view icons to details or vice-versa fixes the issue and let the
item view take the whole space as expected.

Tracked it down to this setSizePolicy() call, since removing that call
doesn't seem to break anything (in KF5 and KF6), then no need to miss with
the widgets size policy.