revise jobuidelegate factorization
- JobUiDelegateFactoryV2 is now extending JobUiDelegateFactor with a
createDelegate that can take a QWidget and flag arguments. this is in
particular to allow us to force creation of delegates to go through the
factory and not ever need to be done directly. this then allows
plasma-integration to supply a different factory and selectively
override delegation with more plasma-esque UIs
- JobUiDelegate existing constructors are deprecated
- new **protected** JobUiDelegate constructors taking a dummy argument
to differentiate them from the deprecated variants (they are protected
such that they aren't called directly anymore - creation must happen
through the factory!)
- all existing of direct construction have been ported to factorization
where the direct construction was not actually required