[KFilePlacesView] Show inline eject button only for removable drives
Otherwise it clutters a list of internal drives that are never meant
to be unmounted, and potentially not even allowed to be, with eject icons.

Also, we currently only show fixed network shares (e.g. through fstab),
which are usually not meant to be unmounted by the user.

Storage without a drive (e.g. a mounted ISO image) keeps its eject inline
button, too.

Triggering an unmount from the contextmenu remains possible (except for
Root and Home, as before).

BUG: 453890