PreviewJob: Resolve parent mime types before checking wildcards
If we have a thumbnailer with a wildcard and the mime type does not exactly match, we would
very likely ignore the parent mime type checks and use the thumbnailer with the wildcard.

In case of the raw-thumbnailer, only the parent mime type of all raw images is registered.
But with the change of the imagethumbnailer to have a wildcard-mimetype, we never call it.

Considering that the wildcards should be used as a fallback, checking it last is reasonable anyways.

BUG: 453480