Pass JobUiDelegate's window to created dialogs
Currently the various dialogs try to use the window associated with a job as their parent window.

However, some/most jobs don't have a window.

If the job doesn't have a window use the window associated with the JobUiDelegate instead

This ensures that the spawned windows are properly parented and e.g. are centered in the main window