Add "Get more Apps in Discover" button to kopenwithdialog
Discover has a neat feature: If you call it with the --mime argument, it shows all
apps that can open this mimetype. This MR adds a button to the "Open With" dialog
that opens discover with list of all apps that can open the selected mimetype. The
button will not be shown when Discover is not installed. This mirrors an equivalent
feature in the XDG portal "Open With" dialog.