KFileWidget: fix infinite recursion after latest changes
In 331aad501ac, the dummy history entry code was removed, which included
blocking signals to m_locationEdit while in setLocationText() methods; this
resulted in infinite recursion, and ultimately a segfault, when selecting
files in the diroperator view.

Thanks to Kai for reporting the crash and doing a git-bisect to find the
guilty commit.