[KUrlNavigatorToggleButton] Load accept icon on demand
There's no need to load that icon until we actually want to start
editing the address bar.

This uncovered a bug in the size calculation on high-dpi where
the `QPixmap::width()` returns device pixels whereas the `sizeHint`
is supposed to be in logical pixels. By returning the hardcoded
value, we ensure the physical size is scaled correctly.