Read .hidden file in file:/ worker rather than KCoreDirLister
This ensure we don't do a blocking file system read and instead
delegate that to the worker which is async.
It also only works for local files anyway.

The downside is that it will likely stop working for other
schemas which return files with a `localPath`, such
as desktop:/

To not jeopardize the performance of "background" list jobs,
such as copy jobs, a new flag `AdditionalHiddenAttributes`
is introduced which will govern whether the ".hidden" (and
NTFS hidden attribute) are read.