Drop obsolete X11 dependency, introduce WITH_X11 option instead
This will only determine whether to search for Qt5X11Extras and then
set HAVE_X11 accordingly. This is a behavior change as previously it was
silently disabled if X11 was not found.

We want to be able to build without X11 support even if some of the used
libraries may not work w/o X11 themselves yet or need to be built with
X11 support for other reverse dependencies.

HAVE_X11 already exists and is set automagically so far, but using
-DCMAKE_DISABLE_FIND_PACKAGE_X11 will break if any dependencies list X11
as required in their cmake config.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>