port http to workerbase
the scariest port yet!

this introduces an unfortunate direct copy of tcpslavebase as
tcpworkerbase.

other than that it is a fairly straight forward port to workerbase and
workerresult, albeit in very complicated code.

workerbase gained an additional setIncomingMetaData function that
replaces the previous direct access of mIncomingMetaData (that is no
longer possible because the member lives on the bridged slavebase inside
the workerbase)