Add DeleteOrTrashJob
KIO::trash()/del() don't ask the user for confirmation, which meant the
caller had to create a UiDelegate of some kind to do that.

This new job basically gathers that logic in one place.

This will replace the code in KDirOperator (and help port away from the
deprecated KIO::JobUiDelegate constructor in Dolphin and Plasma).