kcms/webshortcuts: hide from System Settings' main navigation
This KCM currently lives in the Search group, but is also available as
a sub-window in the Plasma Search KCM, so it's odd to have it apparently
live in two places at once. Let's only show it as a sub-window of the
Plasma Search KCM, which is the place where the user is more likely to
find it anyway, as the only systemwide location where its functionality
appears in KRunner.

Any apps that are using it can embed the KCM in their own settings
window, like Dolphin does for the Trash KCM.